task-queue/requirements.txt
Flask
google-cloud-tasks
gunicorn